ebsc-aws-infrastructure/pyproject.toml
[tool.poetry]
name = "ebsc-aws-infrastructure"
version = "0.1.0"
description = ""
authors = ["Alex Antonison <adantonison@gmal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
aws-cdk-lib = "^2.79.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.267"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
